Fix live power calculation [3/3]
https://wiki-ll.loveliv.es/%E5%88%86%E6%95%B0%E8%AE%A1%E7%AE%97%E6%96%B9%E5%BC%8F

+++ model/live.go
@@ -92,7 +92,7 @@ type LiveScheduleResp struct {
 // Play
 type PlayReq struct {
 	Module           string `json:"module"`
-	PartyUserID      int    `json:"party_user_id"`
+	PartyUserID      int64  `json:"party_user_id"`
 	Action           string `json:"action"`
 	Mgd              int    `json:"mgd"`
 	IsTraining       bool   `json:"is_training"`
